Join and leave party, plus other api bits

+++ api/api/Controllers/PartiesController.cs
@@ -32,18 +32,29 @@ namespace api.Controllers
             _userService = userService;
         }
 
+        [Route("all")]
+        [HttpGet]
+        [Authorize]
+        public async Task<IActionResult> GetAllParties()
+        {
+            return Ok(await _partyService.GetAll());
+        }
+
         // Get user's current party
         [HttpGet]
         [Authorize]
         public async Task<IActionResult> GetParty()
         {
             var userId = User.Claims.Single(c => c.Type == ClaimTypes.PrimarySid).Value;
-            var party = await _context.Users.Where(u => u.Id == userId).Select(u => u.CurrentParty).Include(p => p.Members).FirstOrDefaultAsync();
+            var party = await _partyService.FindByUser(userId);
             
             if (party == null)
             {
                 return NotFound();
             }
+
+            await _context.Entry(party).Collection(p => p.Members).LoadAsync();
+            await _context.Entry(party).Collection(p => p.PendingMembers).LoadAsync();
 
             return Ok(party);
         }
@@ -51,15 +62,14 @@ namespace api.Controllers
         // Create party with current user as owner
         [HttpPost]
         [Authorize]
-        public async Task<IActionResult> CreateParty([FromBody] CreatePartyModel request)
+        public async Task<IActionResult> CreateParty([FromBody] CreatePartyRequest request)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
-            var userId = User.Claims.Single(c => c.Type == ClaimTypes.PrimarySid).Value;
-            var user = await _userService.Find(userId);
+            var user = await GetAuthorizedUser();
             if (user.Owner || user.CurrentParty != null || user.Admin)
             {
                 return StatusCode(StatusCodes.Status400BadRequest, "Already admin or own a party. Delete or leave that party first");
@@ -67,7 +77,7 @@ namespace api.Controllers
             
             try
             {
-                return Ok(await _partyService.Create(userId, request.Name));
+                return Ok(await _partyService.Create(user.Id, request.Name));
             }
             catch (APIException e)
             {
@@ -80,19 +90,90 @@ namespace api.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteParty()
         {
-            var userId = User.Claims.Single(c => c.Type == ClaimTypes.PrimarySid).Value;
-            var user = await _userService.Find(userId);
-            if (user.CurrentParty != null)
+            var user = await GetAuthorizedUser();
+
+            // User is owner of party
+            if (user.Owner && user.CurrentParty != null)
             {
                 await _partyService.Delete(user.CurrentParty.Id);
             }
             return Ok();
         }
 
-
-        private bool PartyExists(string id)
+        [Route("leave")]
+        [HttpPost]
+        [Authorize]
+        public async Task<IActionResult> LeaveParty()
         {
-            return _context.Parties.Any(e => e.Id == id);
+            var user = await GetAuthorizedUser();
+
+            // If user is owner, this is same as deleting the party
+            if (user.CurrentParty != null)
+            {
+                if (user.Owner)
+                {
+                    await _partyService.Delete(user.CurrentParty.Id);
+                }
+                else
+                {
+                    await _partyService.Leave(user.Id);
+                }
+            }
+
+            return Ok();
         }
+
+        // Request to join a party
+        [Route("join")]
+        [HttpPost]
+        [Authorize]
+        public async Task<IActionResult> JoinParty([FromBody] JoinPartyRequest request)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var user = await GetAuthorizedUser();
+            try
+            {
+                await _partyService.RequestToJoin(request.PartyId, user);
+                return Ok();
+            }
+            catch (ResourceNotFoundException e)
+            {
+                return NotFound(e);
+            }
+        }
+
+        // For the user's current party, get the pending members
+        [Route("members/pending")]
+        [HttpGet]
+        [Authorize]
+        public async Task<IActionResult> GetPendingMembers()
+        {
+            var user = await GetAuthorizedUser();
+
+            if (user.CurrentParty != null)
+            {
+                // TODO: make this not include the pendingparty field of each member (all the same)
+                var party = await _partyService.Find(user.CurrentParty.Id);
+                return Ok(user.CurrentParty.PendingMembers);
+            }
+
+            return NotFound();
+        }
+
+        // Helper method to get the user from the current user's auth token. Assumed user is authorized
+        private async Task<User> GetAuthorizedUser()
+        {
+            var userId = User.Claims.Single(c => c.Type == ClaimTypes.PrimarySid).Value;
+            var user = await _userService.Find(userId);
+            await _context.Entry(user).Reference(u => u.CurrentParty).LoadAsync();
+            await _context.Entry(user).Reference(u => u.PendingParty).LoadAsync();
+            return user;
+        }
+
+        
     }
 }
